feat: Add ability to debug wizard pages in dev mode
