+++ src/templates/platform.js
